pyproject.toml
[tool.poetry]
name = "djangosso"
version = "0.0.0"
description = "Django SSO demo"
authors = ["krasnov.v.i <krasnov.v.i@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "3.2"
mypy = "^1.3.0"
black = "^23.3.0"
django-sso = "^3.0.2"
django-cas-ng = "^4.3.0"
djangorestframework-simplejwt = "^5.2.2"
django-simple-sso = "^1.2.0"
PyJWT = "^2.7.0"
cryptography = "^40.0.2"
requests = "^2.31.0"

[tool.poetry.dev-dependencies]
isort = "^5.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
